General update #1
AKA, I did so many different things that it's easier to put it all under
one commit.

Request queueing: If the server has called the API for the same thing
within 2 minutes, it will instead use whatever data it received from the
API the first time. This way, we don't get banned from the API for
excessive use.

Class integration: The classes were written like a Java program. Now,
they are written like a Ruby program. And they actually work. With the
Ruby program. Still a little salty that I had to manually create a
to_json for each of them just for the client though.

Working methods: getTeamMatch, postTeamMatch, saveTeamMatchInfo, and
getSimpleTeamList should all be functioning and fully compatible with
the client.